Leds

+++ Code/Midi/PedalHw/input.h
@@ -6,8 +6,9 @@
 #define PEDALBOARD_INPUT_H
 
 #include <stdbool.h>
+#include <stdint-gcc.h>
 
 int pedal_input_init();
-int pedal_get_current_state(float* volts,bool* stomps);
+int pedal_get_current_state(uint16_t* volts,bool* stomps);
 
 #endif //PEDALBOARD_INPUT_H
